Try and fix old clang on Linux warning

+++ lib/src/plat.c
@@ -209,6 +209,6 @@ void plat_initrandom(void)
     // initialize random number generator
     struct timeval seed;
     gettimeofday(&seed, 0);
-    srandom((seed.tv_sec * 1000) + (seed.tv_usec / 1000));
+    srandom((seed.tv_sec * 1000) + (unsigned int)(seed.tv_usec / 1000));
 #endif
 }
